cleanups and trying to find out what is going out with joint rotations from BVH to TRI..

+++ opengl_acquisition_shared_library/opengl_depth_and_color_renderer/src/Library/ModelLoader/model_loader.c
@@ -843,7 +843,7 @@ int getModelBoneRotationOrderFromBoneName(struct Model *mod,unsigned int boneID)
    {
      if (boneID<triM->header.numberOfBones)
      {
-        return  triM->bones[boneID].info->eulerRotationOrder;
+        return (int)  triM->bones[boneID].info->eulerRotationOrder;
      }
    }
  } else
